Redo handling of watchers
Atlassian roemved the user `name` from the API result.

Instead we are supposed to use the `accountId`.

`accountId` works fine for adding watchers.

But we cannot use it to mention the user inline in a comment (they claim
we should be -- but it doesn't work).

This commit refactors the stuff into using `accountId` for adding the
watchers and the uses the `displayName` in the comment -- but as clear
text, not a mention. That'll have to do now since it is just what's
possible.

To be able to use both `accountId` and `displayName` we need some
refactoring. In this case moving the `\JiraRestApi\User\User` object
around instead of just passing the user name/accoundId around as a
string.